fix: ts error & solve cycle import
