--- console/packages/starwhale-ui/src/Search/constants.ts
@@ -1,19 +1,0 @@
-import { OPERATOR } from '@starwhale/core/datastore'
-
-export enum KIND {
-    BOOLEAN = 'BOOLEAN',
-    CATEGORICAL = 'CATEGORICAL',
-    CUSTOM = 'CUSTOM',
-    DATETIME = 'DATETIME',
-    NUMERICAL = 'NUMERICAL',
-    STRING = 'STRING',
-}
-
-export const FilterTypeOperators: Record<Partial<KIND>, OPERATOR[]> = {
-    [KIND.CATEGORICAL]: [],
-    [KIND.STRING]: [OPERATOR.EQUAL],
-    [KIND.NUMERICAL]: [OPERATOR.EQUAL, OPERATOR.GREATER, OPERATOR.GREATER_EQUAL, OPERATOR.LESS, OPERATOR.LESS_EQUAL],
-    BOOLEAN: [OPERATOR.EQUAL],
-    CUSTOM: [],
-    DATETIME: [],
-}
