+++ console/src/test/provider.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Provider as StyletronProvider } from 'styletron-react'
 import { Client as Styletron } from 'styletron-engine-atomic'
 import { BaseProvider } from 'baseui'
-import themes from '@/theme'
+import deep from '@starwhale/ui/theme'
 import { BrowserRouter } from 'react-router-dom'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import '@/i18n'
@@ -15,7 +15,7 @@ export const WithAll = ({ children }: { children?: React.ReactNode }) => {
     return (
         <QueryClientProvider client={queryClient}>
             <StyletronProvider value={engine}>
-                <BaseProvider theme={themes.deep}>
+                <BaseProvider theme={deep}>
                     <BrowserRouter>{children}</BrowserRouter>
                 </BaseProvider>
             </StyletronProvider>
@@ -24,7 +24,7 @@ export const WithAll = ({ children }: { children?: React.ReactNode }) => {
 }
 
 export function TestBaseProvider({ children }: { children?: React.ReactNode }) {
-    return <BaseProvider theme={themes.deep}>{children}</BaseProvider>
+    return <BaseProvider theme={deep}>{children}</BaseProvider>
 }
 
 export const routeRender = (ui: React.ReactElement, options?: Omit<RenderOptions, 'wrapper'>) =>
