+++ console/vite.config.ts
@@ -8,9 +8,6 @@ import eslint from 'vite-plugin-eslint'
 
 export const alias = {
     'baseui': path.resolve(__dirname, 'node_modules/baseui'),
-    'styletron-react': path.resolve(__dirname, 'node_modules/styletron-react'),
-    'styletron-engine-atomic': path.resolve(__dirname, 'node_modules/styletron-engine-atomic'),
-    'styletron-standard': path.resolve(__dirname, 'node_modules/styletron-standard'),
     '@': path.resolve(__dirname, './src'),
     '@user': path.resolve(__dirname, './src/domain/user'),
     '@project': path.resolve(__dirname, './src/domain/project'),
